refacor code add some more hooks
