+++ client/src/hooks/index.js
@@ -1,3 +1,4 @@
 export { default as useSearchParams } from "./useSearchParams";
 export { usePagination } from "./usePagination";
 export { useFetchTopics } from "./useFetchTopics";
+export { useFetchPublishedResources } from "./useFetchPublishedResources";
